% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_valid}
\alias{cache_valid}
\title{Check cache validity}
\usage{
cache_valid(cache, last_mod_time)
}
\arguments{
\item{cache}{Cached data.}
}
\description{
Check if the cache is newer than the last modification time.
}
