% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input}
\usage{
check_input(index, model, taxon)
}
\arguments{
\item{index}{Character. Update which index?}

\item{model}{Character. Which model to use?}

\item{taxon}{Character. Which taxon?}
}
\description{
Check indicator inputs
}
