% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{check_sp}
\alias{check_sp}
\title{Check species}
\usage{
check_sp(index, sp)
}
\arguments{
\item{index}{Which index to check if species is available for.}

\item{sp}{Which species to check.}
}
\description{
Check if a species is available
}
