% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table.R
\name{create_db_table}
\alias{create_db_table}
\title{Create a DB table}
\usage{
create_db_table(table, ...)
}
\arguments{
\item{table}{Name of table.}

\item{...}{Columns of table.}
}
\description{
Create a database table if it doesn't already exist.
}
