% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{create_input_cache_index}
\alias{create_input_cache_index}
\title{Create an input cache index}
\usage{
create_input_cache_index()
}
\description{
Create a database input cache index if it doesn't already exist.
}
