% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{format_date}
\alias{format_date}
\title{Format date}
\usage{
format_date(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Combine year, month, day of survey into a single date string
}
