% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_from_cache}
\alias{get_from_cache}
\title{Get data from cache}
\usage{
get_from_cache(table, ...)
}
\arguments{
\item{table}{Name of table.}

\item{...}{Column names and data to check in the cache}
}
\description{
Get cached data from a database.
}
