% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{get_from_input_cache}
\alias{get_from_input_cache}
\title{Get data from cache}
\usage{
get_from_input_cache(hash, sp, type)
}
\arguments{
\item{hash}{Hash of input.}

\item{sp}{Species.}

\item{type}{Survey type.}
}
\description{
Get cached data from a database.
}
