% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{get_from_input_cache_index}
\alias{get_from_input_cache_index}
\title{Get data from cache index}
\usage{
get_from_input_cache_index(hash)
}
\arguments{
\item{hash}{Hash of input.}
}
\description{
Get cache index data from a database.
}
