% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_cache.R
\name{get_from_output_cache}
\alias{get_from_output_cache}
\title{Get data from cache}
\usage{
get_from_output_cache(hash)
}
\arguments{
\item{hash}{Hash of input}
}
\description{
Get cached data from a database.
}
