% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices.R
\name{get_indices}
\alias{get_indices}
\title{Get indices}
\usage{
get_indices(file = Sys.getenv("R_CONFIG_FILE"))
}
\arguments{
\item{file}{Configuration file.}
}
\description{
Get indices from a configuration file.
}
