% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Get output}
\usage{
get_output(output, index, model, taxon, region, db)
}
\arguments{
\item{output}{Character. Which type of output?}

\item{index}{Character. Update which index?}

\item{model}{Character. Which model to use?}

\item{taxon}{Character. Which taxon?}

\item{region}{Character. Which region?}

\item{db}{Connection. Database from which to get output.}
}
\description{
Get serialized indicator outputs
}
