% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_svg.R
\name{indicator_svg}
\alias{indicator_svg}
\title{Indicator svg}
\usage{
indicator_svg(year, type)
}
\arguments{
\item{year}{Years to plot.}

\item{type}{Type of indicator.}
}
\description{
Create an svg from an indicator plot.
}
