% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{input_cache_available}
\alias{input_cache_available}
\title{Check if input cache is available}
\usage{
input_cache_available(hash)
}
\arguments{
\item{hash}{Hash of input.}
}
\description{
Check if the cached input data is available.
}
