% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{input_cache_name}
\alias{input_cache_name}
\title{Clean input cache}
\usage{
input_cache_name(...)
}
\arguments{
\item{...}{Name components.}
}
\description{
Remove any unavailable caches from index and any unindexed data.
}
