% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{input_cache_valid}
\alias{input_cache_valid}
\title{Check input cache validity}
\usage{
input_cache_valid(hash, last_mod_time)
}
\arguments{
\item{hash}{Hash of input.}

\item{last_mod_time}{Last time data was modified.}
}
\description{
Check if the input cache is newer than the last modification time.
}
