% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{is_input_cached}
\alias{is_input_cached}
\title{Check if input is cached}
\usage{
is_input_cached(hash)
}
\arguments{
\item{hash}{Hash of input.}
}
\description{
Check if there is any cached input data.
}
