% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_cache.R
\name{output_cache_valid}
\alias{output_cache_valid}
\title{Check cache validity}
\usage{
output_cache_valid(cache, index)
}
\arguments{
\item{cache}{Cached data.}

\item{index}{Which index.}
}
\description{
Check if the cache is newer than the last modification time of the index data.
}
