% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_int_range.R
\name{parse_int_range}
\alias{parse_int_range}
\title{Parse integer range string}
\usage{
parse_int_range(ints)
}
\arguments{
\item{ints}{Integer range as a string.}
}
\description{
Parse a string indicating a range of integers.
}
