% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{pick_first_survey_in_winter}
\alias{pick_first_survey_in_winter}
\title{Pick first survey in winter}
\usage{
pick_first_survey_in_winter(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Pick first winter survey in each year discarding subsequent surveys
}
\details{
This function moves surveys occurring in December ahead one \code{year}.
This enables all December surveys to be grouped with subsequent surveys
occurring in the January of the same winter. Surveys are then grouped by
\code{location_id} and \code{year} and then ordered by date. Then all but the first
survey in each group is removed. If two or more surveys share the same date
and \code{location_id} then one is picked at random and the rest are removed.
This function works on the assumption that surveys are in winter from
December to January and that the \code{surveys} data includes \code{day}, \code{month} and
\code{year} (as integers) and \code{location_id}.
}
