% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{pick_first_survey_in_year}
\alias{pick_first_survey_in_year}
\title{Pick first survey in year}
\usage{
pick_first_survey_in_year(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Pick first survey in each year discarding subsequent surveys
}
