% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_svg.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plot indicator}
\usage{
plot_indicator(year, type)
}
\arguments{
\item{year}{Years to plot.}

\item{type}{Type of indicator.}
}
\description{
Create a plot of an indicator over a range of years.
}
