% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{process_funs}
\alias{process_funs}
\alias{shift_year_winter}
\alias{pick_first_survey_in_winter}
\alias{pick_first_survey_in_year}
\alias{require_seven_fortnights}
\alias{require_gt_two_years}
\alias{pick_first_survey_in_fortnight}
\alias{zero_fill}
\alias{remove_all_zero_locations}
\alias{sum_over_sections}
\alias{sum_by_event}
\alias{set_start_year}
\title{Process functions}
\usage{
process_funs()

shift_year_winter(surveys)

pick_first_survey_in_winter(surveys)

pick_first_survey_in_year(surveys)

require_seven_fortnights(surveys)

require_gt_two_years(surveys)

pick_first_survey_in_fortnight(surveys)

zero_fill(counts, surveys, ...)

remove_all_zero_locations(counts, ...)

sum_over_sections(counts, ...)

sum_by_event(counts, ...)

set_start_year(counts, taxon, ...)
}
\arguments{
\item{surveys}{Survey data.}

\item{counts}{Count data.}

\item{...}{Additional arguments.}

\item{taxon}{Taxon configuration.}
}
\description{
Functions to process indicator input data.
}
