% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{remove_all_zero_locations}
\alias{remove_all_zero_locations}
\title{Remove all-zero locations}
\usage{
remove_all_zero_locations(counts, ...)
}
\arguments{
\item{counts}{Count data.}

\item{...}{Additional arguments.}
}
\description{
Discard locations where taxa always had zero abundance
}
