% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{require_minimum_gaps}
\alias{require_minimum_gaps}
\title{Require minimum gaps}
\usage{
require_minimum_gaps(surveys, ...)
}
\arguments{
\item{surveys}{Survey data.}

\item{...}{Additional arguments.}
}
\description{
Remove survey site-years that too many or to large sampling gaps.
}
\details{
This function groups surveys data by \code{location_id} and \code{year}. It
then removes groups where the survey period has too many or too large
sampling gaps. Where too many is defined as a total gap length over the
\code{year} of 21 days and too large is any single sampling gap of more than
7 days. The function expects the \code{surveys} data to have at least
\code{location_id}, \code{year}, \code{ordinal_day_start} and \code{ordinal_day_end}.
}
