% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{require_minimum_weeks}
\alias{require_minimum_weeks}
\title{Require minimum weeks}
\usage{
require_minimum_weeks(surveys, ...)
}
\arguments{
\item{surveys}{Survey data.}

\item{...}{Additional arguments.}
}
\description{
Remove survey site-years from a region covering less than a minimum number of
weeks.
}
\details{
This function groups surveys data by \code{location_id} and \code{year}. It
then removes groups where the survey period is less than a minimum number
of weeks for a given \code{region}. It expects the \code{surveys} data to have at
least \code{location_id}, \code{year}, \code{region}, \code{ordinal_day_start} and
\code{ordinal_day_end}.
}
