% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{require_two_years}
\alias{require_two_years}
\title{Require at least two years}
\usage{
require_two_years(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Discard locations with less than two survey years
}
