% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_cache}
\alias{set_cache}
\title{Set cache}
\usage{
set_cache(table, ...)
}
\arguments{
\item{table}{Name of table.}

\item{...}{Column names and data to set in the cache.}
}
\description{
Cache data in a database.
}
