% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{set_input_cache}
\alias{set_input_cache}
\title{Set input cache}
\usage{
set_input_cache(name, data, hash, sp = NULL)
}
\arguments{
\item{name}{Name of input.}

\item{data}{Data.}

\item{hash}{Hash of input.}

\item{sp}{Species.}
}
\description{
Cache input data in a database.
}
