% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{set_input_cache_index}
\alias{set_input_cache_index}
\title{Set input cache}
\usage{
set_input_cache_index(name, hash, available)
}
\arguments{
\item{name}{Name of input.}

\item{hash}{Hash of input.}

\item{available}{Is the data available?}
}
\description{
Cache input data in a database.
}
