% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_cache.R
\name{set_output_cache}
\alias{set_output_cache}
\title{Set output cache}
\usage{
set_output_cache(hash, data)
}
\arguments{
\item{hash}{Hash of input}

\item{data}{Raw data.}
}
\description{
Cache output data in a database.
}
