% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{set_start_year}
\alias{set_start_year}
\title{Set start year}
\usage{
set_start_year(counts, taxon, ...)
}
\arguments{
\item{counts}{Count data.}

\item{taxon}{Taxon configuration.}

\item{...}{Additional arguments.}
}
\description{
Discard counts from years before the start year
}
