% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{shift_year_winter}
\alias{shift_year_winter}
\title{Shift year winter}
\usage{
shift_year_winter(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Transpose early winter months to following calendar year
}
