% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_index.R
\name{sp_index}
\alias{sp_index}
\title{Population indicator}
\usage{
sp_index(index, sp, id)
}
\arguments{
\item{index}{Which index.}

\item{sp}{Species.}

\item{id}{Request ID for logging.}
}
\description{
Create a population indicator.
}
