% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species}
\alias{species}
\title{Species list}
\usage{
species(index = "winter_birds", which = c("mx", "spcode"))
}
\arguments{
\item{index}{Which index to get species list from?}

\item{which}{Use MX "mx" codes or species codes "spcode".}
}
\description{
Get species lists for indices
}
