% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{sum_by_event}
\alias{sum_by_event}
\title{Sum by event}
\usage{
sum_by_event(counts, ...)
}
\arguments{
\item{counts}{Count data.}

\item{...}{Additional arguments.}
}
\description{
Sum the counts of a single taxa over the surveys in each year
}
