% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{sum_over_sections}
\alias{sum_over_sections}
\title{Sum over sections}
\usage{
sum_over_sections(counts, ...)
}
\arguments{
\item{counts}{Count data.}

\item{...}{Additional arguments.}
}
\description{
Sum counts over the sections of surveys
}
\details{
This functions groups count data by \code{document_id} (the IDs of the
individual surveys). If multiple taxa \code{counts} are input then data is also
grouped by taxa. Counts are then summed across survey sections when count
data has been provided as surveys split into parts.
}
