% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{sum_taxa_over_sections}
\alias{sum_taxa_over_sections}
\title{Sum taxa over sections}
\usage{
sum_taxa_over_sections(counts, ...)
}
\arguments{
\item{counts}{Count data.}

\item{...}{Additional arguments.}
}
\description{
Sum the counts of multiple taxa over the sections of surveys
}
