% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{svg_ms_index}
\alias{svg_ms_index}
\title{Multi-species index svg}
\usage{
svg_ms_index(index, use_cache, id)
}
\arguments{
\item{index}{Which index?}

\item{use_cache}{Whether to use cached data.}

\item{id}{Request ID for logging.}
}
\description{
Create an svg from a multispecies indicator plot.
}
