% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data}
\usage{
update_data(type, index, taxon, db, do_update = FALSE)
}
\arguments{
\item{type}{Character. Which type of input data (e.g., surveys or counts)}

\item{index}{Character. Update the data of which index?}

\item{taxon}{Character. Update the data for which taxon? Ignored if
\code{type = "surveys"}}

\item{db}{Connection. Database in which to update the data from FinBIF.}

\item{do_update}{Logical. Update data regardless of need.}
}
\description{
Update input data from FinBIF.
}
