% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_index.R
\name{update_index}
\alias{update_index}
\title{Update index}
\usage{
update_index(index, model, db)
}
\arguments{
\item{index}{Character. Update which index?}

\item{model}{Character. Which model to use?}

\item{db}{Connection. Database in which to update index.}
}
\description{
Update index output data.
}
