% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxon_index.R
\name{update_taxon_index}
\alias{update_taxon_index}
\title{Update taxon index}
\usage{
update_taxon_index(index, taxon, db)
}
\arguments{
\item{index}{Character. Update which index?}

\item{taxon}{Character. Update the data for which taxa?}

\item{db}{Connection. Database in which to update index.}
}
\description{
Update the relative abundance index for a taxon.
}
