% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cache.R
\name{wait_for_input_cache_available}
\alias{wait_for_input_cache_available}
\title{Wait for input cache}
\usage{
wait_for_input_cache_available(hash, wait = 10L, time_out = 60L)
}
\arguments{
\item{hash}{Hash of input.}

\item{wait}{How long to wait between checking availability.}

\item{time_out}{How many cycles to wait before timing out.}
}
\description{
Wait for cached input data to become available.
}
