% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{zero_fill}
\alias{zero_fill}
\title{Zero fill}
\usage{
zero_fill(counts, surveys, ...)
}
\arguments{
\item{counts}{Count data.}

\item{surveys}{Survey data.}

\item{...}{Additional arguments.}
}
\description{
Combine count data with survey data filling missing surveys in count data
with zero counts.
}
