% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{combine_with_surveys}
\alias{combine_with_surveys}
\title{Combine with surveys}
\usage{
combine_with_surveys(counts, surveys, ...)
}
\arguments{
\item{counts}{Count data.}

\item{surveys}{Survey data.}

\item{...}{Additional arguments.}
}
\description{
Combine count data with survey data
}
