% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{format_date}
\alias{format_date}
\title{Format date}
\usage{
format_date(surveys, ...)
}
\arguments{
\item{surveys}{Survey data.}

\item{...}{Additional arguments.}
}
\description{
Combine year, month, day of survey into a single date string
}
\details{
This function combines survey \code{year}, \code{month} and \code{day} into a
character string with \code{-} as a separator. The function assumes that survey
data includes \code{year}, \code{month} and \code{day}.
}
