% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_sp_data}
\alias{get_sp_data}
\title{Get species observation data}
\usage{
get_sp_data(index, sp, id)
}
\arguments{
\item{index}{Which index.}

\item{sp}{Species.}

\item{id}{Request ID for logging.}
}
\description{
Get species observation data from finbif.
}
