% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_sp_data}
\alias{get_sp_data}
\title{Get species observation data}
\usage{
get_sp_data(sp, type)
}
\arguments{
\item{sp}{Species.}

\item{type}{Survey type.}
}
\description{
Get species observation data from finbif.
}
