% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get survey data}
\usage{
get_survey_data(name, fltr, slct, id)
}
\arguments{
\item{name}{Name of input.}

\item{fltr}{Filter.}

\item{slct}{Column selection.}

\item{id}{Request ID for logging.}
}
\description{
Get survey data from finbif.
}
