% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_index.R
\name{ms_index}
\alias{ms_index}
\title{Multi-species population indicator}
\usage{
ms_index(index, use_cache, id)
}
\arguments{
\item{index}{Which index.}

\item{use_cache}{Whether to use cached data.}

\item{id}{Request ID for logging.}
}
\description{
Create a multi-species population indicator.
}
