% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_cache.R
\name{output_cache_valid}
\alias{output_cache_valid}
\title{Check cache validity}
\usage{
output_cache_valid(cache, last_mod_time = as.Date("1970-01-01"))
}
\arguments{
\item{cache}{Cached data.}

\item{last_mod_time}{Last time data was modified.}
}
\description{
Check if the cache is newer than the last modification time.
}
