% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{pick_first_survey_in_fortnight}
\alias{pick_first_survey_in_fortnight}
\title{Pick first survey in fortnight}
\usage{
pick_first_survey_in_fortnight(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Pick first survey in each fortnight discarding subsequent surveys
}
