% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{pick_first_survey_in_winter}
\alias{pick_first_survey_in_winter}
\title{Pick first survey in winter}
\usage{
pick_first_survey_in_winter(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Pick first winter survey in each year discarding subsequent surveys
}
