% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{require_seven_fortnights}
\alias{require_seven_fortnights}
\title{Require seven fortnights}
\usage{
require_seven_fortnights(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Divide year into approximate 2 week blocks selecting blocks 11-17 and
discarding locations without a survey in each remaining block
}
