% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{require_two_years}
\alias{require_two_years}
\title{Require at least two years}
\usage{
require_two_years(surveys)
}
\arguments{
\item{surveys}{Survey data.}
}
\description{
Discard locations with less than two survey years
}
\details{
This function groups \code{surveys} by \code{location_id} and then removes all
surveys for locations that do not have data in more than one \code{year}. The
function assumes that \code{surveys} has data for \code{location_id} and \code{year}.
}
