% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_index.R
\name{sp_index}
\alias{sp_index}
\title{Population indicator}
\usage{
sp_index(sp, year, base, id)
}
\arguments{
\item{sp}{Species.}

\item{year}{Year.}

\item{base}{Base year of index.}

\item{id}{Request ID for logging.}
}
\description{
Create a population indicator.
}
