% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{sum_by_event}
\alias{sum_by_event}
\title{Sum by event}
\usage{
sum_by_event(counts, ...)
}
\arguments{
\item{counts}{Count data.}

\item{...}{Additional arguments.}
}
\description{
Sum the counts over the surveys or taxa in each year
}
\details{
This functions groups count data by \code{location_id} and \code{year}. If
multiple taxa \code{counts} are input then data is also grouped by taxa. Counts
are then summed across the survey events at the locations and years.
}
