% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{sum_over_sections}
\alias{sum_over_sections}
\title{Sum over sections}
\usage{
sum_over_sections(counts, ...)
}
\arguments{
\item{counts}{Count data.}

\item{...}{Additional arguments.}
}
\description{
Sum the counts of a single taxa over the sections of surveys
}
