% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{svg_index}
\alias{svg_index}
\title{Index svg}
\usage{
svg_index(sp, year, base)
}
\arguments{
\item{sp}{Species}

\item{year}{Year}

\item{base}{Base year of index}
}
\description{
Create an svg from an indicator plot.
}
