% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{svg_sp_index}
\alias{svg_sp_index}
\title{Species index svg}
\usage{
svg_sp_index(sp, year, base, id)
}
\arguments{
\item{sp}{Species.}

\item{year}{Year.}

\item{base}{Base year of index.}

\item{id}{Request ID for logging.}
}
\description{
Create an svg from an indicator plot.
}
